src/5. Tools and Agents/3_multi_agent_systems/3_Google_A2A_protocol/standalone agents/A2A_standalone_agent_arxiv/agent_tests.py
from dotenv import load_dotenv

from agent import Agent

load_dotenv()

agent = Agent()
sessionID = "session_12345"

query = "Find papers on quantum computing"
response = agent.invoke(query, sessionID)
print(response)
